Widen JIT chunk indices and pack the float_csrmm warp launches
Two follow-ups from sweeping the whole .cu tree for the classes of defect
found in _csr rather than only the files already touched.

Index widening (correctness). The JIT connectivity families index
chunk_counts[row * n_chunks + chunk_id] with both factors as int, so the
product wraps past 2^31 -- the same class as the CSRMM subscripts fixed in
c17f362. n_chunks defaults to 4, but chunk_size is user-settable, so the
product is not bounded by anything the kernel controls. 24 sites across
_jit_{normal,scalar,uniform}/{csr,dt2t}.cu.

Reachability is the same as the CSRMM case and worth stating: the index only
overflows once the indexed buffer itself passes 2^31 elements, i.e. >=8.6 GB.
Out of reach on a small card, reachable on an 80 GB one. The widened multiply
costs nothing -- none of these sites are in an inner loop.

After this pass the only unwidened multiplicative subscripts left tree-wide are
shared-memory strides with a literal * 32, bounded by block size.

Launch shape (performance). binary_csrmm.cu and binary_indexed_csrmm.cu launch
their warp kernels as CSRMM_WARP_RPB rows per 128-thread block, grid-strided
with a capped grid.x. float_csrmm.cu was never updated to match and still used
<<<dim3(m, c_blocks), 32>>>, one warp per block, for both its NT and transpose
warp kernels. Measured on sm_86 (m = k = 16384, n = 64, f32 homo): 2.05x at
avg_nnz 2, 1.58x at 8, 1.28x at 32, 1.19x at 64 -- a win across the entire
range the warp tier serves.

As in the CSRMV change, the row guard becomes `continue` rather than `return`,
since returning inside a grid-strided loop would abandon that warp's remaining
rows.

CSRMM_MAX_GRID_X and CSRMM_WARP_RPB were duplicated verbatim in two .cu files;
rather than add a third copy they move to cuda_common.h next to the CSRMV
geometry.

The spec also records five audits that found nothing, so they are not redone:
shared-memory sizing tree-wide, the binary_fcnmm non-power-of-two warp-count
precondition, dispatch-tier gaps outside Finding 4, warp/barrier convergence,
and the unguarded half atomics.

CSR GPU suite: 715 passed. JIT GPU suites: 1512 passed, 4 skipped.

+++ brainevent/_jit_normal/dt2t.cu
@@ -219,7 +219,7 @@ __global__ void _fill_dt2t_f32_kern(
         seed0, row, chunk_id, lane, cl, chunk_width
     );
     unsigned int lane_offset = warp_exclusive_prefix_u32(lane_count, lane);
-    int base = __ldg(&chunk_offsets[row * n_chunks + chunk_id]) + (int)lane_offset;
+    int base = __ldg(&chunk_offsets[(size_t)row * n_chunks + chunk_id]) + (int)lane_offset;
 
     float loc = __ldg(&w_loc[0]);
     float scale = __ldg(&w_scale[0]);
@@ -355,7 +355,7 @@ __global__ void _fill_dt2t##SFX##_kern( \
         seed0, row, chunk_id, lane, cl, chunk_width \
     ); \
     unsigned int lane_offset = warp_exclusive_prefix_u32(lane_count, lane); \
-    int base = __ldg(&chunk_offsets[row * n_chunks + chunk_id]) + (int)lane_offset; \
+    int base = __ldg(&chunk_offsets[(size_t)row * n_chunks + chunk_id]) + (int)lane_offset; \
     ACC_T loc = READ_W(__ldg(&w_loc[0])); \
     ACC_T scale = READ_W(__ldg(&w_scale[0])); \
     ACC_T y_row = TRANSPOSE ? (ACC_T)0 : READ_W(__ldg(&y[row])); \
